Exec one docker cmd to get the name and status of the containers
