+++ src/memdisk-cloudwatch.go
@@ -31,7 +31,7 @@ func main() {
 }
 
 func reportContainers(icw *instanceCloudwatch) {
-	containers, err := Containers()
+	containers, err := DockeInspect()
 	if err != nil {
 		log.Fatal("Failed to get info of containers:", err)
 	}
